requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
dialogflow==1.1.1
feedparser==6.0.8
Flask==2.0.3
fuzzywuzzy==0.18.0
gnewsclient==1.12
google-api-core==1.31.5
google-auth==1.35.0
googleapis-common-protos==1.56.0
grpcio==1.45.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
packaging==21.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
rsa==4.8
sgmllib3k==1.0.0
six==1.16.0
tornado==6.1
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.9
Werkzeug==2.0.3
